pyproject.toml
[tool.poetry]
name = "avito-tracker"
version = "0.1.0"
description = ""
authors = ["JMURv <architect.lock@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^2.23.1"
psycopg2 = "^2.9.5"
arsenic = "^21.8"
beautifulsoup4 = "^4.11.1"
asyncio = "^3.4.3"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
